+++ R/example.R
@@ -6,7 +6,8 @@
 #' @param name Name of example project. If `NULL`, the example names will be
 #'   listed.
 #' @keywords internal
-exampleDirectory <- function(name = NULL) {
+#' @noRd
+.exampleDirectory <- function(name = NULL) {
   directory <- system.file("extdata", "examples", package = "esqlabsR")
   if (!is.null(name)) {
     directory <- file.path(directory, name)